prototyp interfejsu graficznego

+++ GeneticsAlgorithm/src/Comperators/SortByFitnessComparator.java
@@ -9,6 +9,6 @@ import java.util.Comparator;
 public class SortByFitnessComparator implements Comparator<Individual> {
     @Override
     public int compare(Individual o1, Individual o2) {
-        return Integer.compare(o1.getFitness(), o2.getFitness());
+        return Double.compare(o1.getFitness(), o2.getFitness());
     }
 }
